requirements.txt
influxdb-client==1.49.0
python-dotenv==1.1.1
